Mate: Module 管理

+++ crates/cms-mate/src/domain/form/module.rs
@@ -31,7 +31,7 @@ fn validate_field_icon(ptr: &&String) -> Result<(), ValidationError> {
     validate::string_length(str, false, 0, 30)
 }
 
-fn validate_field_sort(num: i64) -> Result<(), ValidationError> {
+fn validate_field_sort(num: i16) -> Result<(), ValidationError> {
     validate::numeric_range(Some(num), true, 0, 9999)
 }
 
@@ -62,7 +62,7 @@ pub struct ModuleStoreForm {
 
     /// 排序编号
     #[validate(custom(function = "validate_field_sort", message = "排序编号必须在0-9999之间"))]
-    pub sort: Option<i64>,
+    pub sort: Option<i16>,
 
     /// 是否启用
     pub is_enabled: Option<bool>,
